backend/pyproject.toml
[project]
name = "univital-api"
version = "0.1.0"
description = "UniVital Healthcare Financial Risk Engine – backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "hypercorn>=0.15.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "requests",
    "ruff",
]
